DSTORE-1475 add support for conditional upserts
Adds a document processor factory that can be used to conditionall
upsert or skip documents based on the values of specific fields in
either the new or the old documents.

This is mostly intended for handling compliance where the system
handling compliance will not have the original documents available, but
needs to flag documents that may or may not have been already written in
solr.

Example config:

      <lst name="existingPermanentDeletes">
        <str name="must">OLD.compliance_reason:delete</str>
        <str name="action">skip</str>
      </lst>
      <lst name="existingSoftDeletes">
        <str name="must">OLD.compliance_reason:soft_delete</str>
        <str name="action">upsert:compliance_reason</str>
      </lst>
      <lst name="newSoftDeletes">
        <str name="must">NEW.compliance_reason:soft_delete</str>
        <str name="action">upsert:*</str>
      </lst>